[fixed] init wialon jobs

+++ database/migrations/2021_12_10_083844_create_shipment_orders.php
@@ -14,7 +14,7 @@ class CreateShipmentOrders extends Migration
     public function up()
     {
         Schema::create('shipment_orders', function (Blueprint $table) {
-            $table->integer('id')->unique();
+            $table->unsignedBigInteger('id')->unique();
             $table->foreignId('shipment_retail_outlet_id')->constrained()->onDelete('cascade');;
             $table->string('product');
             $table->string('weight')->nullable();
